Refactor routes into package
