+++ routes/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+routes_bp = Blueprint("routes", __name__)
+
+from . import auth, attendance, admin  # noqa: E402,F401
