+++ routes/auth.py
@@ -1,0 +1,23 @@
+from flask import render_template, request, redirect, url_for, flash
+from flask_login import login_user, logout_user, login_required
+from werkzeug.security import check_password_hash
+from model import Uzytkownik
+from . import routes_bp
+
+@routes_bp.route("/login", methods=["GET", "POST"])
+def login():
+    if request.method == "POST":
+        login_val = request.form.get("login")
+        haslo = request.form.get("hasło")
+        user = Uzytkownik.query.filter_by(login=login_val).first()
+        if user and check_password_hash(user.haslo_hash, haslo):
+            login_user(user)
+            return redirect(url_for("routes.admin_dashboard"))
+        flash("Nieprawidłowe dane logowania", "danger")
+    return render_template("login.html")
+
+@routes_bp.route("/logout")
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for("routes.index"))
